Refactor imports to use core directory structure for math, pose, and spline modules
