+++ src/core/math.js
