+++ src/core/spline.js
