+++ src/fly-camera.js
@@ -1,6 +1,6 @@
 import { Vec3 } from 'playcanvas';
 
-import { damp, MyQuat } from './math.js';
+import { damp, MyQuat } from './core/math.js';
 
 const forward = new Vec3();
 const right = new Vec3();
